requirements.txt
# Core data
pandas>=2.1.0
numpy>=1.26.0
pyarrow>=14.0.0
duckdb>=0.9.0
polars>=0.19.0

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0
matplotlib>=3.8.0
contextily>=1.4.0

# Causal inference - classical
statsmodels>=0.14.0
linearmodels>=5.3.0
scipy>=1.11.0

# Causal inference - ML
scikit-learn>=1.3.0
econml>=0.15.0          # DoubleML + CausalForest (GRF-style)
doubleml>=0.7.0         # Alternative DML implementation

# Synthetic DiD
# synthdid: Python port via pysyncon
pysyncon>=0.2.0

# Visualization
plotly>=5.17.0
seaborn>=0.13.0
kaleido>=0.2.1          # plotly static export

# Dashboard
streamlit>=1.28.0
streamlit-folium>=0.15.0
folium>=0.15.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
loguru>=0.7.0
joblib>=1.3.0
requests>=2.31.0
rich>=13.0.0
typer>=0.9.0
